remove fallback to Nouveau
As the user opted in to use the Nvidia driver there is no point in falling
back to Nouveau in case the Nvidia driver didn't load.

The intention of adding the fallback was to make sure the user ends up
with a working desktop, but the framebuffer drivers can provide this as
well and are generally less error prone than Nouveau is. Especially on
more modern GPUs like Pascal there are some known issues resulting in
unstable systems.

As long as no xorg.conf file is provided, X will load a matching DDX, so
that even without loading Nouveau a useable desktop is provided. The sw
accelerated llvmpipe driver will be used to provide OpenGL for software
requiring it.

In case a xorg.conf file was provided, the Nouveau fallback lead to a non
working X server anyway.

Removing the fallback also makes the kmod packages usable on setups where
both drivers are blacklisted, e.g. laptops where Nvidia might be only used
for ondemand graphics offloading or compute workloads and the GPU gets
powered down through other means (e.g. bbswitch).

There is a valid concern about installing Nvidia on systems with
unsupported Nvidia GPUs, but I'd rather have a more noise fallback instead
of a "nothing changes". Maybe it would make sense to have an optional
package to install the Nouveau fallback files?